requirements.txt
# Telegram Bot
aiogram>=3.4.0

# LLM Client (OpenAI-compatible)
openai>=1.12.0

# Database
aiosqlite>=0.19.0

# Scheduler
apscheduler>=3.10.0

# HTTP client
httpx>=0.26.0

# Config
python-dotenv>=1.0.0

# arXiv parser
arxiv>=2.1.0

# HuggingFace Daily Papers RSS
feedparser>=6.0.0

# Utils
pydantic>=2.5.0
